Made some edits to myls for compatability
Added some macro flags to allow compilation on my linux machine.
Probably will not compile on mac anymore.  Going to pull down to one of
the school servers to verify that it will work there and that the
problems I solved are not unique to my own machine.

+++ myls_src/print.c
@@ -19,11 +19,11 @@ void printSCol(FINFO fsi, int entries){
     int row;
     struct passwd *pw;
     struct group *grp;
-    
+
     printf("total %ld\n", fsi.total_blocksize);
     for(row = 0; row < entries; row++){
         printf("%-10.10s  ", sperm(fsi.stat[row].st_mode));
-        printf("%*d ", fsi.max_hardlinks, fsi.stat[row].st_nlink);
+        printf("%*d ", fsi.max_hardlinks, (int)fsi.stat[row].st_nlink);
         pw = getpwuid(fsi.stat[row].st_uid);
         printf("%*s  ", fsi.max_user_name, pw->pw_name);
         grp = getgrgid(fsi.stat[row].st_gid);
@@ -37,21 +37,21 @@ void printSCol(FINFO fsi, int entries){
 /* Multi column print format for listing files not requiring stat information */
 void printCol(struct dirent **ent, const int entries, const int max_name_len){
     int numcols, numrows, row, col, colwidth, base, charCount;
-    
+
     colwidth = (int)(log((double)max_name_len) / log(2));
     colwidth = 8 * (colwidth - 1);
     numcols = (termwidth + colwidth) / colwidth;
     numrows = entries / numcols;
     if(numrows % numcols || numrows == 0)
         numrows++;
-    
+
     for(row = 0; row < numrows; row++){
         for(base = row, col = 0;;){
             charCount =  printf("%s", ent[base]->d_name);
             if((base += numrows) >= entries) break;
             if(++col == numcols) break;
             while(charCount++ < colwidth) (void)putchar(' ');
-            
+
         }
         (void)putchar('\n');
     }
@@ -61,11 +61,11 @@ static void printtime(time_t mtime)
 {
 	int i;
 	char *longstring;
-    
+
 	longstring = ctime(&mtime);
     for (i = 4; i < 11; ++i)
 		(void)putchar(longstring[i]);
-    
+
 #define	SIXMONTHS	((365 / 2) * (24 * 60 * 60))
 	if (mtime + SIXMONTHS > time(NULL))
 		for (i = 11; i < 16; ++i)
